+++ views/diagnosticos/diagnosticos.php
@@ -7,8 +7,8 @@ require_once "views/Inicio.php";
 		<meta charset="utf-8">
 		<meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
 		<title><?php echo $data["titulo"]; ?></title>
-		<link rel="stylesheet" href="assets/css/bootstrap.min.css">
-		<script src="assets/js/bootstrap.min.js" ></script>
+		<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-Zenh87qX5JnK2Jl0vWa8Ck2rdkQ2Bzep5IDxbcnCeuOxjzrPF/et3URy9Bv1WTRi" crossorigin="anonymous">
+   		<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-OERcA2EqjJCMA+/3y+gxIOqMEjwtxJY7qPCqsdltbNJuaOe923+mo//f6V8Qbsw3" crossorigin="anonymous"></script>
 	</head>
 	
 	<body>
@@ -36,7 +36,7 @@ require_once "views/Inicio.php";
 							echo "<tr>";
 							echo "<td>".$dato["diagnostico"]."</td>";
 							echo "<td><a href='index.php?c=diagnosticos&a=modificar&id=".$dato["id_diagnostico"]."' class='btn btn-warning'>Modificar</a></td>";
-							echo "<td><a href='index.php?c=diagnosticos&a=eliminar&id=".$dato["id_diagnostico"]."' class='btn btn-danger'>Eliminar</a></td>";
+							echo "<td><a href='index.php?c=diagnosticos&a=eliminar&id=".$dato["id_diagnostico"]."' class='btn btn-danger' onclick=\"return confirm('Esta seguro de elimiar el registro?');\">Eliminar</a></td>";
 							echo "</tr>";
 						}
 						?>
