+++ views/pacxhabitaciones/pacxhabitaciones.php
@@ -40,7 +40,7 @@ require_once "views/Inicio.php";
 							echo "<td>".$dato["nombre_paciente"]."</td>";
 							echo "<td>".$dato["nombre_habitacion"]."</td>";
 							echo "<td><a href='index.php?c=pacxhabitaciones&a=modificar&id=".$dato["id_pac_x_hab"]."' class='btn btn-warning'>Modificar</a></td>";
-							echo "<td><a href='index.php?c=pacxhabitaciones&a=eliminar&id=".$dato["id_pac_x_hab"]."' class='btn btn-danger'>Eliminar</a></td>";
+							echo "<td><a href='index.php?c=pacxhabitaciones&a=eliminar&id=".$dato["id_pac_x_hab"]."' class='btn btn-danger' onclick=\"return confirm('Esta seguro de elimiar el registro?');\">Eliminar</a></td>";
 							echo "</tr>";
 						}
 						?>
